requirements.txt
beautifulsoup4
certifi
charset-normalizer
idna
lyricsgenius
python-dotenv
redis
requests
six
soupsieve
urllib3
python-vlc
paramiko
pymysql
sshtunnel
dearpygui
pillow
